Use standard boolean type
In C23, bool is a new reserved keyword [1], which means that a custom
enum cannot be named after it anymore. Replace it with a proper bool
type, which fulfills the original purpose of this enum.

Because not all compilers support C23 yet, including MSVC and Apple
Clang [2], include the compatibility header stdbool.h available since
C99 so it builds fine in all platforms.

This change is especially relevant because GCC 15 defaults to C23,
causing freegish to fail to build [3].

[1] https://gcc.gnu.org/gcc-15/porting_to.html#c23-new-keywords
[2] https://en.cppreference.com/w/c/compiler_support/23
[3] https://bugs.debian.org/1096653